Add droidcam-helper script
This script is very specific to my setup and relies on an android phone,
connected via USB, being accessible via adb, and with droidcam software
isntalled on the phone.

The script does nothing more than turning the cellphone screen on, which
seems to be required to start an app; start the droidcam app which will
be serving video via the port 4747 of the device; then start
droidcam-cli app passing `-v adb 4747` to it, which means connect to a
device via adb with video using the port 4747.

Once it's done the cellphone is now webcam device and can be easily used
with BlueJeans, Google Meet, Zoom, etc.

You'll notice the script will be kept alive and once you hit ^C it'll
clean itself up by killing all the droidcam-cli processes and stopping
the app inside the cellphone.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>